backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
firebase_admin>=6.0.0
google-generativeai>=0.8.0
langgraph>=0.2.0
langchain-core>=0.3.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
stripe>=8.0.0
